сейв

+++ hw/maksim_lamaka/lesson04/lecture.py
@@ -1,6 +1,7 @@
 from decimal import Decimal
 from typing import Any
 
+
 def task_01_money(rubles: int, coins: int, amount: int) -> Decimal:
     rubles = Decimal(rubles)
     coins = Decimal(coins)
@@ -8,7 +9,7 @@ def task_01_money(rubles: int, coins: int, amount: int) -> Decimal:
     return (rubles + (coins / 100)) * amount
 
 
-def task_02_sign(number: Any) -> int:
+def task_02_sign(number: Any) -> float:
     if isinstance(number, complex):
         return 0
     elif number > 0:
